requirements.txt
anthropic
python-dotenv
pyyaml
graphviz
groq
pyperclip
pytest